pyproject.toml
[project]
name = "python-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "couchbase>=4.3.2",
    "flask>=3.0.3",
    "flask-restx>=1.3.0",
    "prometheus-client>=0.21.0",
    "python-dotenv>=1.0.1",
]
